refactor: move to runtime common
