+++ runtimes/common/src/lib.rs
@@ -47,6 +47,7 @@ pub mod constants;
 pub mod fees;
 pub mod migrations;
 pub mod pallet_id;
+pub mod xcm_config;
 
 #[cfg(feature = "runtime-benchmarks")]
 pub mod benchmarks;
