+++ runtimes/peregrine/Cargo.toml
@@ -137,6 +137,7 @@ runtime-benchmarks = [
   "parachain-staking/runtime-benchmarks",
   "rococo-runtime/runtime-benchmarks",
   "sp-runtime/runtime-benchmarks",
+  "xcm-builder/runtime-benchmarks",
 ]
 std = [
   "attestation/std",
