+++ runtimes/spiritnet/Cargo.toml
@@ -136,6 +136,7 @@ runtime-benchmarks = [
   "parachain-staking/runtime-benchmarks",
   "rococo-runtime/runtime-benchmarks",
   "sp-runtime/runtime-benchmarks",
+  "xcm-builder/runtime-benchmarks",
 ]
 std = [
   "attestation/std",
